requirements.txt
fastapi~=0.115.12
SQLAlchemy~=2.0.40
python-jose~=3.4.0
passlib~=1.7.4
python-dotenv~=1.1.0
pydantic~=2.11.3
openai~=1.76.0
pandas~=2.2.3
joblib~=1.4.2
scikit-learn~=1.6.1
xgboost~=3.0.0
pydub~=0.25.1
Levenshtein~=0.27.1